requirements.txt
# Core ML libraries
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Computer Vision
opencv-python>=4.5.0
Pillow>=8.3.0

# Deep Learning
tensorflow>=2.8.0
# atau pytorch jika menggunakan PyTorch
# torch>=1.10.0
# torchvision>=0.11.0

# Web framework
Flask>=2.0.0
Flask-CORS>=3.0.0

# Progress bars
tqdm>=4.62.0

# File handling
pickle-mixin>=1.0.2

# Image processing tambahan
imageio>=2.9.0
scikit-image>=0.18.0

# Utility
pathlib2>=2.3.0
argparse>=1.4.0